elastic search implementation

+++ src/Domain/Entities/Shared/AttributeValueModel.cs
@@ -7,7 +7,7 @@ public class AttributeValueModel
     [JsonPropertyName("itemIds")] public List<string>? ItemIds { get; set; }
     
     [JsonPropertyName("stringValue")] public List<TranslationModel>? StringValue { get; set; }
-    [JsonPropertyName("numericValue")] public decimal? NumericValue { get; set; }
+    [JsonPropertyName("numericValue")] public double? NumericValue { get; set; }
 
     [JsonPropertyName("numericRangeValue")]
     public AttributeValueNumericRangeModel? NumericRangeValue { get; set; }
